chore(suite): sort tokens also in send form
